Refactor color filename
